move tests from quantity to ecmetadata to avoid circular deps
