+++ core/quantity/package.json
@@ -34,10 +34,8 @@
     "url": "http://www.bentley.com"
   },
   "devDependencies": {
-    "@bentley/units-schema": "^1.0.9",
     "@itwin/build-tools": "workspace:*",
     "@itwin/core-bentley": "workspace:*",
-    "@itwin/ecschema-metadata": "workspace:*",
     "@itwin/eslint-plugin": "5.2.2-dev.2",
     "@types/glob": "^5.0.35",
     "@vitest/coverage-v8": "^3.0.6",
