+++ core/quantity/src/test/Formatting.test.ts
@@ -2,24 +2,12 @@
 * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
-import * as path from "path";
-import * as fs from "fs";
-import { beforeAll, describe, expect, it } from "vitest";
-import { DOMParser } from "@xmldom/xmldom";
+import { describe, expect, it } from "vitest";
 import { QuantityError } from "../Exception";
 import { Format } from "../Formatter/Format";
 import { DecimalPrecision, FormatTraits, parseFormatTrait } from "../Formatter/FormatEnums";
 import { CustomFormatProps, FormatProps } from "../Formatter/Interfaces";
-import { FormatterSpec } from "../Formatter/FormatterSpec";
 import { TestUnitsProvider } from "./TestUtils/TestHelper";
-import { Schema, SchemaContext, SchemaReadHelper, SchemaUnitProvider, XmlParser } from "@itwin/ecschema-metadata";
-
-function deserializeXmlSync(schemaXml: string, context: SchemaContext) {
-  const parser = new DOMParser();
-  const document = parser.parseFromString(schemaXml);
-  const reader = new SchemaReadHelper(XmlParser, context);
-  return reader.readSchemaSync(new Schema(context), document);
-}
 
 describe("Formatting tests:", () => {
   it("A valid 'string' Type is required", async () => {
@@ -1236,189 +1224,4 @@ describe("Formatting tests:", () => {
   });
 });
 
-describe("Formatting tests with more extensive units:", () => {
-  let context: SchemaContext;
-  let provider: SchemaUnitProvider;
 
-  beforeAll(() => {
-    context = new SchemaContext();
-
-    const schemaFile = path.join(__dirname, "..", "..", "node_modules", "@bentley", "units-schema", "Units.ecschema.xml");
-    const schemaXml = fs.readFileSync(schemaFile, "utf-8");
-    deserializeXmlSync(schemaXml, context);
-
-    provider = new SchemaUnitProvider(context);
-  });
-
-  it("should format Kelvin to Fahrenheit with negative result", async () => {
-    const format = await Format.createFromJSON("Test", provider, {
-      type: "Decimal",
-      formatTraits: ["ShowUnitLabel"],
-      composite: {
-        units: [
-          {
-            name: "Units.FAHRENHEIT",
-          },
-        ],
-      },
-    });
-    const persistenceUnit = await provider.findUnitByName("Units.K");
-    const formatterSpec = await FormatterSpec.create("Test", format, provider, persistenceUnit);
-    const formatted = formatterSpec.applyFormatting(100);
-    expect(formatted).to.eql("-279.67 °F");
-  });
-
-  it("should format Celsius to Fahrenheit with negative result", async () => {
-    const format = await Format.createFromJSON("Test", provider, {
-      type: "Decimal",
-      formatTraits: ["ShowUnitLabel", "trailZeroes"],
-      precision: 2,
-      composite: {
-        units: [
-          {
-            name: "Units.FAHRENHEIT",
-          },
-        ],
-      },
-    });
-    const persistenceUnit = await provider.findUnitByName("Units.CELSIUS");
-    const formatterSpec = await FormatterSpec.create("Test", format, provider, persistenceUnit);
-    const formatted = formatterSpec.applyFormatting(-40);
-    expect(formatted).to.eql("-40.00 °F");
-  });
-
-  it("should format Fahrenheit to Celsius with negative result", async () => {
-    const format = await Format.createFromJSON("Test", provider, {
-      type: "Decimal",
-      formatTraits: ["ShowUnitLabel"],
-      precision: 2,
-      composite: {
-        units: [
-          {
-            name: "Units.CELSIUS",
-          },
-        ],
-      },
-    });
-    const persistenceUnit = await provider.findUnitByName("Units.FAHRENHEIT");
-    const formatterSpec = await FormatterSpec.create("Test", format, provider, persistenceUnit);
-    const formatted = formatterSpec.applyFormatting(0);
-    expect(formatted).to.eql("-17.78 °C");
-  });
-
-  it("should format Rankine to Celsius with negative result", async () => {
-    const format = await Format.createFromJSON("Test", provider, {
-      type: "Decimal",
-      formatTraits: ["ShowUnitLabel"],
-      precision: 2,
-      composite: {
-        units: [
-          {
-            name: "Units.CELSIUS",
-          },
-        ],
-      },
-    });
-    const persistenceUnit = await provider.findUnitByName("Units.RANKINE");
-    const formatterSpec = await FormatterSpec.create("Test", format, provider, persistenceUnit);
-    const formatted = formatterSpec.applyFormatting(200);
-    expect(formatted).to.eql("-162.04 °C");
-  });
-
-  it("should handle temperature conversion with fractional format", async () => {
-    const format = await Format.createFromJSON("Test", provider, {
-      type: "Fractional",
-      formatTraits: ["ShowUnitLabel"],
-      precision: 8,
-      composite: {
-        units: [
-          {
-            name: "Units.FAHRENHEIT",
-          },
-        ],
-      },
-    });
-    const persistenceUnit = await provider.findUnitByName("Units.K");
-    const formatterSpec = await FormatterSpec.create("Test", format, provider, persistenceUnit);
-    const formatted = formatterSpec.applyFormatting(150);
-    expect(formatted).to.eql("-189 5/8 °F");
-  });
-
-  it("should handle zero Kelvin (absolute zero) conversion", async () => {
-    const format = await Format.createFromJSON("Test", provider, {
-      type: "Decimal",
-      formatTraits: ["ShowUnitLabel"],
-      precision: 2,
-      composite: {
-        units: [
-          {
-            name: "Units.FAHRENHEIT",
-          },
-        ],
-      },
-    });
-    const persistenceUnit = await provider.findUnitByName("Units.K");
-    const formatterSpec = await FormatterSpec.create("Test", format, provider, persistenceUnit);
-    const formatted = formatterSpec.applyFormatting(0);
-    expect(formatted).to.eql("-459.67 °F");
-  });
-
-  it("should handle positive temperature conversion correctly", async () => {
-    const format = await Format.createFromJSON("Test", provider, {
-      type: "Decimal",
-      formatTraits: ["ShowUnitLabel"],
-      precision: 2,
-      composite: {
-        units: [
-          {
-            name: "Units.FAHRENHEIT",
-          },
-        ],
-      },
-    });
-    const persistenceUnit = await provider.findUnitByName("Units.K");
-    const formatterSpec = await FormatterSpec.create("Test", format, provider, persistenceUnit);
-    const formatted = formatterSpec.applyFormatting(373.15);
-    expect(formatted).to.eql("212 °F");
-  });
-
-  it("should convert negative Celsius to positive Kelvin", async () => {
-    const format = await Format.createFromJSON("Test", provider, {
-      type: "Decimal",
-      formatTraits: ["ShowUnitLabel"],
-      precision: 2,
-      composite: {
-        units: [
-          {
-            name: "Units.K",
-            label: "K"
-          },
-        ],
-      },
-    });
-    const persistenceUnit = await provider.findUnitByName("Units.CELSIUS");
-    const formatterSpec = await FormatterSpec.create("Test", format, provider, persistenceUnit);
-    const formatted = formatterSpec.applyFormatting(-100);
-    expect(formatted).to.eql("173.15 K");
-  });
-
-  it("should convert negative Fahrenheit to positive Rankine", async () => {
-    const format = await Format.createFromJSON("Test", provider, {
-      type: "Decimal",
-      formatTraits: ["ShowUnitLabel"],
-      precision: 2,
-      composite: {
-        units: [
-          {
-            name: "Units.RANKINE",
-          },
-        ],
-      },
-    });
-    const persistenceUnit = await provider.findUnitByName("Units.FAHRENHEIT");
-    const formatterSpec = await FormatterSpec.create("Test", format, provider, persistenceUnit);
-    const formatted = formatterSpec.applyFormatting(-200);
-    expect(formatted).to.eql("259.67 °R");
-  });
-
-});
